replace `delimiter` with `get_iter` in config
Using `delimiter` is not ideal because it's not very flexible.  For instance,
using a delimiter of ':' doesn't work great when grep results contain colons:

  f.ts:12:function foo(a: string) {

Ideally, the user could pass a full pattern match, such as '^([^:]-):(.*)$'.
However, lua's string.match does not return match indices, so we can't adjust
match positions to be highlighted.

This is the next best thing i can think of.